testing

+++ routes/notes.js
@@ -6,12 +6,39 @@ const router = express.Router();
 const mongoose = require('mongoose');
 
 const Note = require('../models/note');
+const Folder = require('../models/folder');
+const Tag = require('../models/tag');
+
+function validateFolderId(userId, folderId) {
+  if (!folderId) {
+    return Promise.resolve();
+  }
+  return Folder.findOne({ _id: folderId, userId })
+    .then(result => {
+      if (!result) {
+        return Promise.reject('InvalidFolder');
+      }
+    });
+}
+
+function validateTagIds(userId, tags = []) {
+  if (!tags.length) {
+    return Promise.resolve();
+  }
+  return Tag.find({ $and: [{ _id: { $in: tags }, userId }] })
+    .then(results => {
+      if (tags.length !== results.length) {
+        return Promise.reject('InvalidTag');
+      }
+    });
+}
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/notes', (req, res, next) => {
   const { searchTerm, folderId, tagId } = req.query;
+  const userId = req.user.id;
 
-  let filter = {};
+  let filter = { userId };
 
   /**
    * BONUS CHALLENGE - Search both title and content using $OR Operator
@@ -23,13 +50,9 @@ router.get('/notes', (req, res, next) => {
     filter.title = { $regex: re };
   }
 
-  if (folderId) {
-    filter.folderId = folderId;
-  }
+  if (folderId) filter.folderId = folderId;
 
-  if (tagId) {
-    filter.tags = tagId;
-  }
+  if (tagId) filter.tags = tagId;
 
   Note.find(filter)
     .populate('tags')
@@ -45,6 +68,7 @@ router.get('/notes', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -52,7 +76,8 @@ router.get('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  Note.findById(id)
+  Note.findById({ _id: id, userId })
+    .select('title content created folderId tags')
     .populate('tags')
     .then(result => {
       if (result) {
@@ -69,6 +94,9 @@ router.get('/notes/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/notes', (req, res, next) => {
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
+
+  const newItem = { title, content, tags, userId };
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -87,13 +115,27 @@ router.post('/notes', (req, res, next) => {
     });
   }
 
-  const newItem = { title, content, folderId, tags };
+  if (mongoose.Types.ObjectId.isValid(folderId)) {
+    newItem.folderId = folderId;
+  }
 
-  Note.create(newItem)
+  const valFolderIdProm = validateFolderId(userId, folderId);
+  const valTagIdsProm = validateTagIds(userId, tags);
+
+  Promise.all([valFolderIdProm, valTagIdsProm])
+    .then(() => Note.create(newItem))
     .then(result => {
       res.location(`${req.originalUrl}/${result.id}`).status(201).json(result);
     })
     .catch(err => {
+      if (err === 'InvalidFolder') {
+        err = new Error('The folder is not valid');
+        err.status = 400;
+      }
+      if (err === 'InvalidTag') {
+        err = new Error('The tag is not valid');
+        err.status = 400;
+      }
       next(err);
     });
 });
@@ -102,6 +144,7 @@ router.post('/notes', (req, res, next) => {
 router.put('/notes/:id', (req, res, next) => {
   const { id } = req.params;
   const { title, content, folderId, tags } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!title) {
@@ -116,7 +159,7 @@ router.put('/notes/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateItem = { title, content, tags };
+  const updateItem = { title, content, tags, userId };
 
   if (mongoose.Types.ObjectId.isValid(folderId)) {
     updateItem.folderId = folderId;
@@ -132,10 +175,13 @@ router.put('/notes/:id', (req, res, next) => {
     });
   }
 
-  const options = { new: true };
+  const valFolderIdProm = validateFolderId(userId, folderId);
+  const valTagIdsProm = validateTagIds(userId, tags);
 
-  Note.findByIdAndUpdate(id, updateItem, options)
-    .populate('tags')
+  Promise.all([valFolderIdProm, valTagIdsProm])
+    .then(() => {
+      return Note.findByIdAndUpdate(id, updateItem, { new: true }).populate('tags');
+    })
     .then(result => {
       if (result) {
         res.json(result);
@@ -144,6 +190,14 @@ router.put('/notes/:id', (req, res, next) => {
       }
     })
     .catch(err => {
+      if (err === 'InvalidFolder') {
+        err = new Error('The folder is not valid');
+        err.status = 400;
+      }
+      if (err === 'InvalidTag') {
+        err = new Error('The tag is not valid');
+        err.status = 400;
+      }
       next(err);
     });
 });
@@ -151,10 +205,11 @@ router.put('/notes/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/notes/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
-  Note.findByIdAndRemove(id)
-    .then(() => {
-      res.status(204).end();
+  Note.findByIdAndRemove({ _id: id, userId })
+    .then(result => {
+      result ? res.status(204).end() : next();
     })
     .catch(err => {
       next(err);
