+++ routes/tags.js
@@ -10,7 +10,10 @@ const Note = require('../models/note');
 
 /* ========== GET/READ ALL ITEMS ========== */
 router.get('/tags', (req, res, next) => {
-  Tag.find()
+
+  const userId = req.user.id;
+
+  Tag.find({ userId })
     .sort('name')
     .then(results => {
       res.json(results);
@@ -23,6 +26,7 @@ router.get('/tags', (req, res, next) => {
 /* ========== GET/READ A SINGLE ITEM ========== */
 router.get('/tags/:id', (req, res, next) => {
   const { id } = req.params;
+  const userId = req.user.id;
 
   if (!mongoose.Types.ObjectId.isValid(id)) {
     const err = new Error('The `id` is not valid');
@@ -30,7 +34,7 @@ router.get('/tags/:id', (req, res, next) => {
     return next(err);
   }
 
-  Tag.findById(id)
+  Tag.findById({ _id: id, userId })
     .then(result => {
       if (result) {
         res.json(result);
@@ -46,8 +50,9 @@ router.get('/tags/:id', (req, res, next) => {
 /* ========== POST/CREATE AN ITEM ========== */
 router.post('/tags', (req, res, next) => {
   const { name } = req.body;
+  const userId = req.user.id;
 
-  const newTag = { name };
+  const newTag = { name, userId };
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -73,6 +78,7 @@ router.post('/tags', (req, res, next) => {
 router.put('/tags/:id', (req, res, next) => {
   const { id } = req.params;
   const { name } = req.body;
+  const userId = req.user.id;
 
   /***** Never trust users - validate input *****/
   if (!name) {
@@ -87,7 +93,7 @@ router.put('/tags/:id', (req, res, next) => {
     return next(err);
   }
 
-  const updateTag = { name };
+  const updateTag = { name, userId };
 
   Tag.findByIdAndUpdate(id, updateTag, { new: true })
     .then(result => {
@@ -109,7 +115,9 @@ router.put('/tags/:id', (req, res, next) => {
 /* ========== DELETE/REMOVE A SINGLE ITEM ========== */
 router.delete('/tags/:id', (req, res, next) => {
   const { id } = req.params;
-  const tagRemovePromise = Tag.findByIdAndRemove(id);
+  const userId = req.user.id;
+
+  const tagRemovePromise = Tag.findByIdAndRemove({ _id: id, userId });
   // const tagRemovePromise = Tag.remove({ _id: id }); // NOTE **underscore** _id
 
   const noteUpdatePromise = Note.updateMany(
@@ -126,7 +134,8 @@ router.delete('/tags/:id', (req, res, next) => {
       }
     })
     .catch(err => {
-      next(err); });
+      next(err);
+    });
 
 });
